+++ src/main/java/com/lib/library_management/Entity/BooksEntity.java
@@ -36,7 +36,7 @@ public class BooksEntity {
     @Transient
     String SubjectCategory;
     @Transient
-    Integer BookCode;
+    Long BookCode;
     @Transient
     String Author;
     @Transient
